refactor(storage): refactor
